requirements.txt
Django==5.2.11
djangorestframework==3.16.1
stripe>=14.3.0
python-dotenv==1.2.1